Add files via upload
# First Python Data Analysis Project
This project represents my first experience of using Python to analyze data. After a three-month course starting with the basics, the task was to analyze COVID-19 data using the COVID-19 Vaccine Data Repository made available by the European Centre for Disease Prevention and Control (ECDC) along with any other dataset.

The goal of the assignment was to showcase my ability to use the `pandas` package to create meaningful visualizations and insights into the COVID-19 phenomenon.

## Key Takeaways
- Familiarity with basic Python syntax and data analysis libraries
- Understanding of how to import, clean and manipulate data using `pandas`
- Ability to create visualizations to represent and interpret data insights

In my project, I decided to focus initially on analysing the spread of the pandemic in different parts of the world in order to identify patterns that might be of interest and on which I could focus.
Subsequently, narrowing the scope of my research, I studied the differences in the implementation of the vaccination campaign in three European countries that differed in the type of political ideology in government at the time; and then concluded with a very rudimentary study of what could be the linear effect of the spread of vaccination on some variables of interest in the three different countries.